+++ Integrations/Anomali_ThreatStream_v2/Anomali_ThreatStream_v2.py
@@ -382,54 +382,41 @@ def test_module():
     demisto.results('ok')
 
 
+def ips_reputation_command(ip, threshold=None, status="active,inactive"):
+    ips = argToList(ip, ',')
+    for single_ip in ips:
+        get_ip_reputation(single_ip, threshold, status)
+
+
 def get_ip_reputation(ip, threshold=None, status="active,inactive"):
     """
         Checks the reputation of given ip from ThreatStream and
         returns the indicator with highest severity score.
     """
-
     params = build_params(value=ip, type="ip", status=status, limit=0)
     indicator = search_indicator_by_params(params, ip)
     threshold = threshold or DEFAULT_THRESHOLD
+    dbot_context = get_dbot_context(indicator, threshold)
+    ip_context = get_ip_context(indicator, threshold)
+    threat_ip_context = get_threat_generic_context(indicator)
 
-    indicator = {
-        'value': '8.8.8.8',
-        'type': 'ip',
-        'asn': 'asn1',
-        'country': 'US',
-        'latitude': 1000,
-        'longitude': 1000,
-        'confidence': 80,
-        'org': 'Demisto',
-        'status': 'active',
-        'meta': {
-            'severity': 'high'
-        }
+    ec = {
+        'DBotScore': dbot_context,
+        'IP(val.Address == obj.Address)': ip_context,
+        'ThreatStream.IP(val.Address == obj.Address)': threat_ip_context
     }
-    indicator_severity = indicator['meta']['severity']
-    dbot_score = {
-        'high': DBotScore.BAD,
-        'medium': DBotScore.SUSPICIOUS,
-        'low': DBotScore.GOOD
-    }[indicator_severity]
+    human_readable = tableToMarkdown(F"IP reputation for: {ip}", threat_ip_context)
 
-    ip_context = IP(
-        ip=ip,
-        asn=indicator.get('asn'),
-        geo_country=indicator.get('country'),
-        geo_latitude=indicator.get('latitude'),
-        geo_longitude=indicator.get('longitude'),
-        dbot_score=dbot_score
-    )
-    command_result = CommandResult(
-        vendor='ThreatStream',
-        object_name='IP',
-        uniq_field='Address',
-        output=indicator,
-        indicators=ip_context
-    )
+    return_outputs(human_readable, ec, indicator)
 
-    return command_result
+
+def domains_reputation_command(domain, threshold=None, status="active,inactive"):
+    """
+        Wrapper function for get_url_reputation.
+    """
+    domains = argToList(domain, ',')
+    for single_domain in domains:
+        get_domain_reputation(single_domain, threshold, status)
 
 
 def get_domain_reputation(domain, threshold=None, status="active,inactive"):
@@ -440,51 +427,27 @@ def get_domain_reputation(domain, threshold=None, status="active,inactive"):
     params = build_params(value=domain, type="domain", status=status, limit=0)
     indicator = search_indicator_by_params(params, domain)
     threshold = threshold or DEFAULT_THRESHOLD
+    dbot_context = get_dbot_context(indicator, threshold)
+    domain_context = get_domain_context(indicator, threshold)
+    threat_domain_context = get_threat_generic_context(indicator)
 
-    indicator = {
-        'value': 'demisto.com',
-        'ip': '8.8.8.8',
-        'type': 'domain',
-        'created_ts': '2019-01-01T00:00:00',
-        'modified_ts': '2019-01-01T00:00:00',
-        'confidence': 80,
-        'org': 'Demisto',
-        'status': 'active',
-        'meta': {
-            'severity': 'low',
-            'registrant_name': 'Demisto',
-            'registrant_email': 'info@demisto.com',
-            'registrant_phone': '050-0000111'
-        }
+    ec = {
+        'DBotScore': dbot_context,
+        'Domain(val.Name == obj.Name)': domain_context,
+        'ThreatStream.Domain(val.Address == obj.Address)': threat_domain_context
     }
-    indicator_severity = indicator['meta']['severity']
-    dbot_score = {
-        'high': DBotScore.BAD,
-        'medium': DBotScore.SUSPICIOUS,
-        'low': DBotScore.GOOD
-    }[indicator_severity]
+    human_readable = tableToMarkdown(F"Domain reputation for: {domain}", threat_domain_context)
 
-    domain_indicator = Domain(
-        name=indicator.get('value'),
-        ip=indicator.get('ip'),
-        whois=Whois(
-            registrant=Registrant(
-                email=indicator.get('meta').get('registrant_email'),
-                name=indicator.get('meta').get('registrant_name'),
-                phone=indicator.get('meta').get('registrant_phone')
-            )
-        ),
-        dbot_score=dbot_score,
-        original_indicator=indicator
-    )
-    command_result = CommandResult(
-        vendor='ThreatStream',
-        object_name='Analysis',
-        uniq_field='value',
-        output=domain_indicator
-    )
+    return_outputs(human_readable, ec, indicator)
 
-    return command_result
+
+def files_reputation_command(file, threshold=None, status="active,inactive"):
+    """
+        Wrapper function for get_url_reputation.
+    """
+    files = argToList(file, ',')
+    for single_file in files:
+        get_file_reputation(single_file, threshold, status)
 
 
 def get_file_reputation(file, threshold=None, status="active,inactive"):
@@ -511,6 +474,15 @@ def get_file_reputation(file, threshold=None, status="active,inactive"):
     human_readable = tableToMarkdown(F"MD5 reputation for: {file}", threat_file_context)
 
     return_outputs(human_readable, ec, indicator)
+
+
+def urls_reputation_command(url, threshold=None, status="active,inactive"):
+    """
+        Wrapper function for get_url_reputation.
+    """
+    urls = argToList(url, ',')
+    for single_url in urls:
+        get_url_reputation(single_url, threshold, status)
 
 
 def get_url_reputation(url, threshold=None, status="active,inactive"):
@@ -583,17 +555,29 @@ def get_passive_dns(value, type="ip", limit=50):
 
 
 def import_ioc_with_approval(import_type, import_value, confidence="50", classification="Private",
-                             threat_type="exploit", severity="low"):
+                             threat_type="exploit", severity="low", ip_mapping=False, domain_mapping=False,
+                             url_mapping=False, email_mapping=False, md5_mapping=False):
     """
         Imports indicators data to ThreatStream.
         The data can be imported using one of three import_types: data-text (plain-text),
         file-id of uploaded file to war room or URL.
     """
+    ip_mapping = demisto.args().get('ip_mapping', 'no') == 'yes'
+    domain_mapping = demisto.args().get('domain_mapping', 'no') == 'yes'
+    url_mapping = demisto.args().get('url_mapping', 'no') == 'yes'
+    email_mapping = demisto.args().get('email_mapping', 'no') == 'yes'
+    md5_mapping = demisto.args().get('md5_mapping', 'no') == 'yes'
+
     files = None
     uploaded_file = None
     data = {
         'confidence': confidence,
         'classification': classification,
+        'ip_mapping': ip_mapping,
+        'domain_mapping': domain_mapping,
+        'url_mapping': url_mapping,
+        'email_mapping': email_mapping,
+        'md5_mapping': md5_mapping,
         'threat_type': threat_type,
         'severity': severity
     }
@@ -607,10 +591,15 @@ def import_ioc_with_approval(import_type, import_value, confidence="50", classif
 
         uploaded_file = open(file_info['path'], 'rb')
         files = {'file': (file_info['name'], uploaded_file)}
+        params = build_params()
     else:
-        data[import_type] = import_value
+        if import_value == 'url':
+            params = build_params(url=import_value)
+        else:
+            params = build_params(datatext=import_value)
+
     # in case import_type is not file-id, http_requests will receive None as files
-    res = http_request("POST", "v1/intelligence/import/", params=CREDENTIALS, data=data, files=files)
+    res = http_request("GET", "v1/intelligence/import/", params=params, data=data, files=files)
     # closing the opened file if exist
     if uploaded_file:
         uploaded_file.close()
@@ -865,13 +854,13 @@ def main():
         if demisto.command() == 'test-module':
             test_module()
         elif demisto.command() == 'ip':
-            demisto.results(get_ip_reputation(**args))
+            ips_reputation_command(**args)
         elif demisto.command() == 'domain':
-            get_domain_reputation(**args)
+            domains_reputation_command(**args)
         elif demisto.command() == 'file':
-            get_file_reputation(**args)
+            files_reputation_command(**args)
         elif demisto.command() == 'url':
-            get_url_reputation(**args)
+            urls_reputation_command(**args)
         elif demisto.command() == 'threatstream-email-reputation':
             get_email_reputation(**args)
         elif demisto.command() == 'threatstream-get-passive-dns':
@@ -908,20 +897,6 @@ def main():
             return_error("The server is not reachable.")
         else:
             return_error(e)
-
-
-class CommandResult:
-    def __init__(self):
-        self.output = None
-        self.dbot_score = None
-        self.readable_output = ''
-        self.raw_response = None
-
-    def add_output(self, output):
-        self.output = output
-
-    def set_readable_output(self, markdown):
-        self.readable_output = markdown
 
 
 # python2 uses __builtin__ python3 uses builtins