+++ Scripts/SumList/CHANGELOG.md
@@ -1,5 +1,5 @@
 ## [Unreleased]
-
+Fixed an issue with handling input as comma separated string.
 
 ## [19.8.2] - 2019-08-22
 #### New Script