+++ Scripts/SumList/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## [Unreleased]
-Fixed an issue with handling input as comma separated string.
+- Fixed an issue with handling input as comma separated string.
+- Added support to floating numbers.
 
 ## [19.8.2] - 2019-08-22
 #### New Script